+++ sdmx-dl-provider-ri/src/main/java/module-info.java
@@ -1,9 +1,10 @@
-import sdmxdl.provider.ri.spi.VaultService;
+import internal.util.credentials.SwingPasswordPrompt;
+import internal.util.credentials.WindowsPasswordPrompt;
 import internal.util.credentials.WindowsVaultService;
 import sdmxdl.file.spi.FileCaching;
 import sdmxdl.file.spi.Reader;
 import sdmxdl.provider.ri.authenticators.MsalAuthenticator;
-import sdmxdl.provider.ri.authenticators.WinPasswordVaultAuthenticator;
+import sdmxdl.provider.ri.authenticators.BasicAuthenticator;
 import sdmxdl.provider.ri.caching.RiCaching;
 import sdmxdl.provider.ri.drivers.FileRiDriver;
 import sdmxdl.provider.ri.drivers.RngRiDriver;
@@ -13,6 +14,8 @@ import sdmxdl.provider.ri.monitors.UptimeRobotMonitor;
 import sdmxdl.provider.ri.networking.RiNetworking;
 import sdmxdl.provider.ri.readers.XmlReader;
 import sdmxdl.provider.ri.registry.RiRegistry;
+import sdmxdl.provider.ri.spi.PasswordPrompt;
+import sdmxdl.provider.ri.spi.VaultService;
 import sdmxdl.web.spi.*;
 
 module sdmxdl.provider.ri {
@@ -34,6 +37,7 @@ module sdmxdl.provider.ri {
     requires nbbrd.io.http;
     requires nbbrd.net.proxy;
     requires com.microsoft.aad.msal4j;
+    requires java.desktop;
 
     exports sdmxdl.provider.ri.drivers to sdmxdl.provider.dialects, sdmxdl.provider.px;
     exports sdmxdl.provider.ri.spi;
@@ -48,7 +52,7 @@ module sdmxdl.provider.ri {
 
     provides Authenticator with
             MsalAuthenticator,
-            WinPasswordVaultAuthenticator;
+            BasicAuthenticator;
 
     provides Monitor with
             UpptimeMonitor,
@@ -66,9 +70,14 @@ module sdmxdl.provider.ri {
     provides Registry with
             RiRegistry;
 
+    provides PasswordPrompt with
+            SwingPasswordPrompt,
+            WindowsPasswordPrompt;
+
     provides VaultService with
             WindowsVaultService;
 
+    uses PasswordPrompt;
     uses VaultService;
 
     opens sdmxdl.provider.ri.monitors to com.google.gson;