+++ sdmx-dl-provider-ri/src/test/java/sdmxdl/provider/ri/caching/VaultCacheTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import sdmxdl.web.Credentials;
 import tests.sdmxdl.ext.FakeClock;
 
-import java.net.PasswordAuthentication;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.HashMap;
@@ -35,7 +34,7 @@ class VaultCacheTest {
 
     @Test
     public void testGet() {
-        Map<String, PasswordAuthentication> map = new HashMap<>();
+        Map<MockedVaultService.Key, String> map = new HashMap<>();
         FakeClock clock = new FakeClock();
         Duration ttl = Duration.ofSeconds(1);
 
@@ -52,24 +51,30 @@ class VaultCacheTest {
                 .as("Empty map should return null")
                 .isNull();
 
-        PasswordAuthentication r1000 = new PasswordAuthentication("r1", "r1".toCharArray());
-        map.put("KEY1", r1000);
+        MockedVaultService.Key key1 = new MockedVaultService.Key("KEY1", "KEY1");
+
+        String r1000 = "r1";
+        map.put(key1, r1000);
         clock.set(1009);
         assertThat(x.get("KEY1"))
                 .as("Existing key should return value")
-                .isEqualTo(Credentials.of(r1000, clock.instant().plus(ttl)));
+                .returns("KEY1", credentials -> credentials != null ? credentials.getCredentials().getUserName() : null)
+                .returns(r1000.toCharArray(), credentials -> credentials != null ? credentials.getCredentials().getPassword() : null)
+                .returns(clock.instant().plus(ttl), credentials -> credentials != null ? credentials.getExpirationTime() : null);
 
         clock.set(1009);
         assertThat(x.get("KEY2"))
                 .as("Non-existing key should return null")
                 .isNull();
 
-        PasswordAuthentication r1009 = new PasswordAuthentication("r2", "r2".toCharArray());
-        map.put("KEY1", r1009);
+        String r1009 = "r2";
+        map.put(key1, r1009);
         clock.set(1010);
         assertThat(x.get("KEY1"))
                 .as("Updated key should return updated value")
-                .isEqualTo(Credentials.of(r1009, clock.instant().plus(ttl)));
+                .returns("KEY1", credentials -> credentials != null ? credentials.getCredentials().getUserName() : null)
+                .returns(r1009.toCharArray(), credentials -> credentials != null ? credentials.getCredentials().getPassword() : null)
+                .returns(clock.instant().plus(ttl), credentials -> credentials != null ? credentials.getExpirationTime() : null);
     }
 
     private static @NonNull Condition<@Nullable Credentials> emptyCredentials(Instant expirationTime) {